Review tests

+++ posts/views.py
@@ -10,7 +10,7 @@ from django.views.decorators.cache import cache_page
 @cache_page(20, key_prefix="index_page")
 def index(request):
     post_list = Post.objects.select_related(
-        'author', 'group'
+        "author", "group"
     ).order_by("-pub_date").all()
 
     paginator = Paginator(post_list, 10)
@@ -47,7 +47,6 @@ def add_comment(request, username, post_id):
     return redirect(url)
 
 
-# @cache_page(60 * 15)
 def group_posts(request, slug):
     group = get_object_or_404(Group, slug=slug)
     posts = group.group_posts.all()
@@ -66,14 +65,12 @@ def group_posts(request, slug):
 
 @login_required
 def new_post(request):
-    if request.method == "POST":
-        form = PostForm(request.POST, files=request.FILES)
-        if form.is_valid():
-            form.instance.author = request.user
-            form.save()
-            url = reverse('index')
-            return redirect(url)
-    form = PostForm()
+    form = PostForm(request.POST or None, files=request.FILES or None)
+    if request.method == "POST" and form.is_valid():
+        form.instance.author = request.user
+        form.save()
+        url = reverse("index")
+        return redirect(url)
     labels = {
         "title": "Новая запись",
         "button": "Добавить новую запись"
@@ -83,23 +80,6 @@ def new_post(request):
     )
 
 
-def profile_extract(username):
-    """ Get data for profile template """
-    author = get_object_or_404(User, username=username)
-    # Note that we swap followers and following
-    # because it is not ours profile but author's
-    following = author.follower.count()
-    followers = author.following.count()
-    posts_count = author.posts.count()
-    return {
-        "author": author,
-        "followers": followers,
-        "following": following,
-        "posts_count": posts_count
-    }
-
-
-# @cache_page(60 * 15)
 def profile(request, username):
     author = get_object_or_404(User, username=username)
     posts = author.posts.all()
@@ -118,25 +98,22 @@ def profile(request, username):
         {
             "page": page,
             "paginator": paginator,
-            "profile": profile_extract(username),
             "author": author,
-            'following': following
+            "following": following
         }
     )
 
 
-# @cache_page(60 * 15)
 def post_view(request, username, post_id):
     post = get_object_or_404(Post, pk=post_id, author__username=username)
-    author = get_object_or_404(User, username=username)  # pass test
-    comments = post.post_comments.all()
+    author = post.author
+    comments = post.comments.all()
     form = CommentForm()
     return render(
         request, "posts/post.html",
         {
             "post": post,
-            "profile": profile_extract(username),
-            "username": author,  # pass test
+            "author": author,
             "form": form,
             "items": comments
         }
@@ -196,14 +173,12 @@ def server_error(request):
 @login_required
 def follow_index(request):
     logged = request.user
-    followings = Follow.objects.select_related(
-        'author', 'user'
-    ).filter(user=logged)
 
-    posts = []
-    for following in followings:
-        author = User.objects.get(username=following.author)
-        posts += author.posts.all()
+    # 1. Get author object of post
+    # 2. Get posts "following" of author via related name
+    # 3. Where user = logged.
+    posts = Post.objects.filter(author__following__user=logged)
+
     paginator = Paginator(posts, 10)
     page_number = request.GET.get("page")
     page = paginator.get_page(page_number)
@@ -212,17 +187,17 @@ def follow_index(request):
         {
             "page": page,
             "paginator": paginator,
-            'profile': profile_extract(logged)
+            "username": logged
         }
     )
 
 
 @login_required
 def profile_follow(request, username):
-    url = reverse('profile', args=[username])
+    url = reverse("profile", args=[username])
     user = request.user
     author = get_object_or_404(User, username=username)
-    if Follow.objects.filter(user=user, author=author):
+    if Follow.objects.filter(user=user, author=author).exists():
         return redirect(url)
     if author == user:
         return redirect(url)
@@ -234,9 +209,10 @@ def profile_follow(request, username):
 
 @login_required
 def profile_unfollow(request, username):
-    url = reverse('profile', args=[username])
+    url = reverse("profile", args=[username])
     user = request.user
     author = get_object_or_404(User, username=username)
-    instance = get_object_or_404(Follow, user=user, author__exact=author)
-    instance.delete()
+    if Follow.objects.filter(user=user, author=author).exists():
+        instance = Follow.objects.filter(user=user, author=author)
+        instance.delete()
     return redirect(url)
